Metodo de Entrada

+++ Banco/01-Alterações.sql
@@ -3,7 +3,7 @@ CREATE TABLE cliente (
                          categoria_saldo varchar(100),
                          descricao_saldo varchar(100),
                          data_recebimento varchar(100),
-                         saldo_extra integer,
+                         saldo_extra numeric,
                          CONSTRAINT cliente_pk PRIMARY KEY (id)
 );
